Refactors CharacterSheet into multiple components, uses reusable CharacterSheetBox component
