+++ src/components/characterSheet/CharacterSheetBox.tsx
@@ -1,0 +1,66 @@
+import React, { FC, useState } from 'react';
+import { Box } from 'grommet';
+import { FormUp, FormDown, Edit } from 'grommet-icons';
+
+import { HeadingWS } from '../../config/grommetConfig';
+import { useFonts } from '../../contexts/fontContext';
+
+interface CharacterSheetBoxProps {
+  title: string;
+  children: JSX.Element;
+  open?: boolean;
+  navigateToCharacterCreation?: (step: string) => void;
+  targetCreationStep?: string;
+}
+
+const CharacterSheetBox: FC<CharacterSheetBoxProps> = ({
+  title,
+  children,
+  open,
+  navigateToCharacterCreation,
+  targetCreationStep,
+}) => {
+  const [show, setShow] = useState(open);
+  const { crustReady } = useFonts();
+
+  const toggleShow = () => setShow(!show);
+  return (
+    <Box
+      fill="horizontal"
+      align="center"
+      justify="start"
+      pad={{ vertical: '12px' }}
+      style={{ borderBottom: '1px solid rgba(255, 255, 255, 0.25)' }}
+    >
+      <Box fill="horizontal" direction="row" align="center" justify="between" pad={{ vertical: '12px' }}>
+        <HeadingWS
+          crustReady={crustReady}
+          level="3"
+          margin="0px"
+          alignSelf="start"
+          onClick={toggleShow}
+          style={{ cursor: 'pointer' }}
+        >
+          {title}
+        </HeadingWS>
+        <Box direction="row" align="center" gap="12px">
+          {show ? (
+            <FormUp onClick={toggleShow} style={{ cursor: 'pointer' }} />
+          ) : (
+            <FormDown onClick={toggleShow} style={{ cursor: 'pointer' }} />
+          )}
+          {!!navigateToCharacterCreation && targetCreationStep && (
+            <Edit
+              color="accent-1"
+              onClick={() => navigateToCharacterCreation(targetCreationStep)}
+              style={{ cursor: 'pointer' }}
+            />
+          )}
+        </Box>
+      </Box>
+      {show && children}
+    </Box>
+  );
+};
+
+export default CharacterSheetBox;
